git-svn-id: https://svn.cv.nrao.edu/svn/casa/release/current@11107 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/tableplot/implement/TablePlot/BasePlot.cc
@@ -3533,7 +3533,7 @@ Int BasePlot::flagData(Int direction, String msname, String spwexpr,
                    }
                }
 
-               if (row < 0 || row >= selRows) {
+               if (row < 0 || row >= (Int)selRows) {
                   continue;
                }
